+++ frontend/src/components/Results/Criterias/CriteriaRequests.vue
@@ -27,7 +27,8 @@
                         <td class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25">
                             {{ request.method }}
                         </td>
-                        <td class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25 text-left pl-4">
+                        <td
+                            class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25 text-left break-all pl-4">
                             <a :href="request.url" class="hover:underline text-clip">
                                 {{ request.url }}
                             </a>