+++ frontend/src/view/AboutPage.vue
@@ -1,5 +1,5 @@
 <template>
-    <BasicCard class="m-16">
+    <BasicCard class="m-2 md:m-16">
         <h1 class="text-gray-800 dark:text-gray-200 text-center text-4xl mb-4 font-bold">About</h1>
     </BasicCard>
 </template>